+++ src/real_time_cv/file_utils.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+import tempfile
+
+
+def save_file(uploaded_file):
+    temp_file = tempfile.NamedTemporaryFile(delete=False)
+    temp_file.write(uploaded_file.getbuffer())
+    filename = temp_file.name
+    return filename
